Complete game

+++ src/Model/Ship.cs
@@ -38,7 +38,7 @@ namespace MyGame
             _tiles = new List<Tile>();
 
             // gets the ship size from the enumarator
-            _sizeOfShip = _shipName;
+            _sizeOfShip = (int) _shipName;
         }
 
         public int Column
